Refactor btree_index.rs: Split 1,241-line file into focused modules
Addresses issue #1604 by splitting the monolithic btree_index.rs file into
smaller, focused modules organized by responsibility.

## Changes

### New Module Structure
- `btree_index.rs` (293 lines) - Core struct, constructors, metadata management
- `operations/` - Tree-level operations by concern:
  - `node_ops.rs` (113 lines) - Basic node-level insert/delete/search
  - `insert.rs` (182 lines) - Insertion, split propagation, root creation
  - `delete.rs` (154 lines) - Deletion entry points
  - `rebalance.rs` (264 lines) - Rebalancing, borrowing, merging after deletions
  - `query.rs` (194 lines) - Point lookups, range scans, multi-key lookups
- `bulk_load.rs` (169 lines) - Bottom-up bulk loading
- `traversal.rs` (100 lines) - Tree navigation helpers

### Benefits
- Main file reduced from 1,241 to 293 lines (76% reduction)
- Clear separation of concerns (CRUD operations isolated)
- Each module under 300 lines for easier comprehension
- Improved code navigation with logical grouping
- All 147 existing tests pass without modification
- No changes to public API

## Testing
- ✅ All 147 tests in vibesql-storage pass
- ✅ No changes to public API or behavior
- ✅ Code compiles without warnings
