+++ crates/vibesql-storage/src/btree/node/operations/mod.rs
@@ -1,0 +1,9 @@
+//! B+ tree operations modules
+//!
+//! This module organizes all B+ tree operations into focused submodules.
+
+pub(crate) mod node_ops;  // Basic node-level operations (must be first)
+pub(crate) mod insert;
+pub(crate) mod delete;
+pub(crate) mod rebalance;
+pub(crate) mod query;
